oracle-eye/requirements.txt
telethon
sqlmodel
sqlalchemy
python-dotenv
asyncio
